Fixed Incrementation Bug & Disabled Buttons on Play
Fixed a bug where the incrementation would not calculate the current value since the denomation carried over from the previous selection.

Additionally, disabled denomination buttons at the beginning of the bonus round.

Finally, cleared out the currentWins for both Text and value so that the number of wins isn't too astronomically high.