Fixed Chest Swapper Function
Chest Swapper function

Sorts through and swaps a 0 value chest with a value higher than 0. THis way no matter what chest the player picks, the chest gives the the values they won.

From there, it checks to make sure that if all chests remaining are 0, then to end the round and make it the "Pooper"