Balance Bug
There has been an issue where checking the balance causes a doubling of the starting amount.

The fix was to prevent the same value from being added to the balance twice.

Bug found that when player takes away an amount with the denomicaiton, it does not add it back with BalanceFix in the increment script if that denomination is the same value as the balance.